+++ lib/Net/Z3950/FOLIO/Config.pm
@@ -189,19 +189,21 @@ Net::Z3950::FOLIO::Config - configuration file for the FOLIO Z39.50 gateway
 
 =head1 DESCRIPTION
 
-The FOLIO Z39.50 gateway C<z2folio> is configured by a single file,
-named on the command-line, and expressed in JSON.  This file specifies
-how to connect to FOLIO, how to log in, and how to translate its
-instance records into MARC.
+The FOLIO Z39.50 gateway C<z2folio> is configured by a stacking set of
+JSON files whose basename is specified on the command-line. These
+files specify how to connect to FOLIO, how to log in, and how to
+search.
 
-The structure of the file is pretty simple. There are several
-top-level section, each described in its own section below, and each
-of them an object with several keys that can exist in it.
+The structure of each of these file is the same, and the mechanism by
+which they are stacked is described below. The shared format is
+simple. There are several top-level sections, each described in its own
+section below, and each of them is an object with several keys that can
+exist in it.
 
 If any string value contains sequences of the form C<${NAME}>, they
 are each replaced by the values of the corresponding environment
 variables C<$NAME>, providing a mechanism for injecting values into
-the condfiguration. This is useful if, for example, it is necessary to
+the configuration. This is useful if, for example, it is necessary to
 avoid embedding authentication secrets in the configuration file.
 
 When substituting environment variables, the bash-like fallback syntax
@@ -329,6 +331,53 @@ every query submitted by the client, so it acts as a filter allowing
 through only records that satisfy it. This might be used, for example,
 to specify C<source=marc> to limit search result to only to those
 FOLIO instance records that were translated from MARC imports.
+
+=head1 CONFIGURATION STACKING
+
+To implement both multi-tenancy and per-database output tweaks that may be required for specific Z39.50 client application, it is necessary to allow flexibility in the configuration of the server, based both on tenant and on further specifications. Three levels of configuration are supported.
+
+=over 4
+
+=item 1
+
+A base configuration file is always used, and will typically provide the bulk of the configuraiton that is the same for all supported tenants and filters. Its base name is specified when the server is invoked -- for example, as the argument to the C<-c> command-line option of C<z2folio>: when the server is invoked as C<z2folio -c etc/config> the base configuration file is found at C<etc/config.json>.
+
+=item 2
+
+The Z39.50 database name provided in a search is used as a name of a sub-configuration specific to that database, and is also passed to FOLIO as the name of the tenant to be addressed. So for example, if a Z93.50 search request come in for the database C<theo>, then the additional tenant-specific configuration file C<etc/config.theo.json> is also consulted. It is acceptable for the file not to exist, in which case there is no tenant-specific configuration and the base configuration is used for that database.
+
+Values provided in a tenant-specific configuration are added to those of the base configuration, overriding the base values when the same item is provided at both levels.
+
+=item 3
+
+One or more I<filters> may also be used to specify additional configuration. These are specified as part of the Z39.50 database name, separated from the main part of the name by pipe characters (C<|>). For example, if the database name C<theo|foo|bar> is used, the two additional filter-specific configuration files are also read, C<etc/config.foo.json> and C<etc/config.bar.json>. The values of filter-specific configurations override those of the base or tenant-specific configuration, and those of later filters override those of earlier filters.
+
+=back
+
+In the example used here, then, a server launched with C<z2folio -c etc/config> and serving a search against the Z39.50 database C<theo|foo|bar> will consult four configuration files:
+C<etc/config.json>,
+C<etc/config.theo.json> (if present),
+C<etc/config.foo.json>
+and
+C<etc/config.bar.json>.
+
+This scheme allows us to handle several scenarios in a uniform way:
+
+=over 4
+
+=item *
+
+Basic configuration all in one place
+
+=item *
+
+Tenant-specific overrides, such as the customer-specific definition of ISBN searching (see issue ZF-24) in the tenant configuration, leaving the standard definition to apply to other tenants.
+
+=item *
+
+Application-specific overrides, such as those needed by the ABLE client (see issue ZF-25), all specific only in a filter that is not used except when explicitly requested.
+
+=back
 
 =head1 SEE ALSO